Added more files for refactoring and arranged minishell.h in order before refactoring tokenizer
